beaconboard: initial watchdog time increased. mainboard: lot of bugs fixed. At least seems to work fine!! secondary_robot: some bugs fixed.

+++ beaconboard/main.h
@@ -89,7 +89,7 @@ struct beaconboard
 
 	/* beacon running watchdog */
 	uint8_t watchdog_enable;
-	uint8_t watchdog;
+	uint16_t watchdog;
 #define WATCHDOG_NB_TIMES	10
 
 };
